requirements.txt
pandas
boto3
awswrangler
great_expectations
streamlit
plotly
pyathena
python-dotenv
dbt-core
dbt-athena-community